+++ tests/static-analysis/stubs/autodescription.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace {
+    function the_seo_framework() {}
+}
+
+namespace The_SEO_Framework\Builders {
+    class Sitemap_Base {
+        /**
+         * @param string $sitemap_id
+         *
+         * @return string
+         */
+        public function generate_sitemap($sitemap_id = 'base') {}
+    }
+}